requirements.txt
pandas
matplotlib.pyplot
json
seaborn
numpy 
matplotlib.cm